--- PawCut/PawCut/Sources/Views/PawcutCamera/PawcutTip.swift
@@ -1,114 +1,0 @@
-//
-//  PawcutTip.swift
-//  PawCut
-//
-//  Created by 광로 on 7/17/25.
-//
-
-import SwiftUI
-
-struct PawcutTip: View {
-    @Environment(\.dismiss) private var dismiss
-    // TODO: 현재는 dismiss만 처리 중.NavigationStack으로 관리 예정
-    
-    var body: some View {
-        NavigationView {
-            ZStack {
-                Color.grayScale06
-                    .ignoresSafeArea()
-                
-                VStack(spacing: 0) {
-                    // 상단바
-                    HStack {
-                        Button(action: {
-                            dismiss()
-                        }) {
-                            Image(systemName: "chevron.left")
-                                .font(.system(size: 17, weight: .semibold))
-                                .foregroundColor(.grayScale01)
-                        }
-                        .frame(width: 33, height: 44)
-                        
-                        Spacer()
-                    }
-                    ScrollView {
-                        VStack(alignment: .leading, spacing: 0) {
-                            // 타이틀
-                            HStack {
-                                Text("촬영 전에 확인해주세요!")
-                                    .pretendardFont(size: ._24, weight: .bold)
-                                    .foregroundColor(.grayScale01)
-                                    .padding(.top, 56)
-                                    .padding(.leading, 20)
-                                
-                                Spacer()
-                            }
-                            // 리스트
-                            let tips: [(icon: String, text: String)] = [
-                                ("tip_camera", "'촬영하기'를 누르면 바로 촬영이 시작돼요."),
-                                ("tip_clock", "6초 타이머가 자동으로 작동해요."),
-                                ("tip_picture", "사진은 기본으로 8장이 연속 촬영돼요."),
-                                ("tip_hand", "설정에서 카메라 접근을 먼저 허용해 주세요."),
-                                ("tip_arrow", "뒤로가기 버튼으로 촬영을 중단할 수 있어요."),
-                                ("camera_sound", "음성을 통해 강아지의 시선을 ~~~~")
-                            ]
-                            VStack(spacing: 25) {
-                                ForEach(tips, id: \.icon) { tip in
-                                    TipRow(iconName: tip.icon, text: tip.text)
-                                }
-                            }
-                            .padding(.top, 60)
-                            .padding(.horizontal, 32)
-                            
-                            
-                        }
-                    }
-                    // TODO: 버튼 동작 (PawcutReady로 이동 및 추후 수정 예정)
-                    VStack(spacing: 0) {
-                        NavigationLink(destination: PawcutReady()) {
-                            Text("촬영하기")
-                                .pretendardFont(size: ._16, weight: .semibold)
-                                .foregroundColor(.grayScale06)
-                                .frame(maxWidth: .infinity)
-                                .frame(height: 59)
-                                .background(
-                                    RoundedRectangle(cornerRadius: 8)
-                                        .fill(.pointPurple01)
-                                )
-                        }
-                        .padding(.horizontal, 20)
-                        .padding(.bottom, 20)
-                    }
-                    .background(.grayScale06)
-                }
-            }
-            
-        }
-        
-    }
-}
-
-struct TipRow: View {
-    let iconName: String
-    let text: String
-    
-    var body: some View {
-        HStack(alignment: .top, spacing: 12) {
-            Image(iconName)
-                .resizable()
-                .aspectRatio(contentMode: .fit)
-                .frame(width: 20, height: 20)
-            
-            Text(text)
-                .pretendardFont(size: ._16, weight: .semibold)
-                .foregroundColor(.grayScale02)
-                .fixedSize(horizontal: false, vertical: true)
-            
-            Spacer()
-        }
-    }
-}
-
-#Preview {
-    PawcutTip()
-}
